route file and services file reorganized
